+++ src/components/InstructorLessonsCard/index.js
@@ -1,9 +1,9 @@
 import React from 'react'
-import Widget from 'components/Widget'
+import Card from 'components/Card'
 import LessonListsByStates from 'components/LessonListsByStates'
 
 export default ({instructor}) => (
-  <Widget title={`${instructor.first_name}'s lessons`}>
+  <Card title={`${instructor.first_name}'s lessons`}>
     <LessonListsByStates instructor={instructor} />
-  </Widget>
+  </Card>
 )
