+++ src/screens/Dashboard/components/GetPublishedCard/components/Checklist/components/MoreInfo.js
