+++ src/screens/Dashboard/components/GetPublishedCard/utils/isStepComplete/index.js
