+++ src/screens/Dashboard/components/GetPublishedCard/utils/isStepComplete/index.test.js
