+++ src/screens/Dashboard/components/InstructorRevenueCard/utils/removeRevenueMonth/index.js
