+++ src/screens/Dashboard/components/InstructorRevenueCard/utils/removeRevenueMonth/index.test.js
