+++ src/screens/Instructors/screens/Instructor/components/InstructorInfoCard/components/InfoBit/index.js
